requirements.txt
wheel
psycopg2-binary
python-dotenv
questionary
chess